Ready for layout
Opted to keep 4 trampoline/led controls. I ended up using some of the LED control lines as also control lines to the 7-segment, because I won't need to use both of these at the same time.

The toddler trampline is the only one that uses the 7-segment, and it will only use 2 buttons, so button 5,6,7 can be used for 7-segment.

Also added in 12V barrel jack input for 7-segment.

-added ethernet connector for control cable to overhead lights. These IO will control the beefcake relay boards that live up in the rafters with the can lights.