memory now has global timeout to free jump
it will repeat memory round until global timeout happens.

note, it's a little longer than GLOBAL_STATIC_TIMEOUT_LIMIT because, it needs to fail a memory game to eventually check that global static timeout. This is okay, it just means that you need to let it sit for 2 fail cycles for it to finally get back to free jump